Update src/cxtracker.c
Added ability to roll off disk when free disk space is less than n bytes.

Added ability to store pcaps to disk in directories named YYYY-MM-DD.

Note: Roll-off is currently recursive, in case we want to extend directory tree later or the user wants to group directories by week / month / etc in a modification, outside script, or by hand.

It looks at directories alphabetically but the files by age.  I'm not sure if I'm fine with the inconsistency or not.  Initial feeling is yes, because user may want to alter the directories.